requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Data providers
yfinance>=0.2.0
alpha-vantage>=2.3.0
python-binance>=1.0.0
ccxt>=4.0.0

# Technical analysis
ta-lib>=0.4.0
pandas-ta>=0.3.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Statistical analysis
statsmodels>=0.13.0
scikit-learn>=1.0.0

# Configuration management
pyyaml>=6.0
python-dotenv>=0.19.0

# Database
sqlalchemy>=1.4.0
sqlite3

# Logging
loguru>=0.6.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.0.0

# Development tools
black>=22.0.0
flake8>=5.0.0
mypy>=0.991

# Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.0.0

# Performance
numba>=0.56.0
joblib>=1.1.0

# API clients
requests>=2.28.0
websocket-client>=1.4.0

# Time series analysis
arch>=5.3.0
